application/requirements.txt
streamlit
openai
pdfplumber
python-docx